Keep doctor available before native guard
