+++ src/cli-main.ts
@@ -14,14 +14,7 @@ import {
   SettingsManager,
 } from "@earendil-works/pi-coding-agent";
 import { loadEnv } from "./config.js";
-import { renderDoctorReport } from "./doctor/render.js";
-import { buildDoctorReport, type DoctorModelSetupState } from "./doctor/report.js";
-import { getProvider, type ProviderId } from "./onboarding/providers.js";
-import {
-  clearProviderOnboardingEntry,
-  loadOnboardingState,
-  saveOnboardingState,
-} from "./onboarding/state.js";
+import { handleDoctorCommand } from "./doctor/cli-command.js";
 import { createOpenCandleSession } from "./pi/session.js";
 import { continueOpenCandleSession } from "./pi/session-storage.js";
 
@@ -163,74 +156,6 @@ async function handleMonitorCommand(args: string[], cwd: string): Promise<boolea
   });
   process.exitCode = exitCode;
   return true;
-}
-
-async function handleDoctorCommand(
-  args: string[],
-  cwd: string,
-  agentDir: string,
-): Promise<boolean> {
-  if (args[0] !== "doctor") return false;
-
-  const json = args.includes("--json");
-  const enableFlag = args.findIndex((arg) => arg === "--enable" || arg === "--reenable");
-  if (enableFlag >= 0) {
-    const providerId = args[enableFlag + 1] as ProviderId | undefined;
-    if (!providerId) {
-      console.error("Usage: opencandle doctor --enable <provider>");
-      process.exitCode = 1;
-      return true;
-    }
-    try {
-      getProvider(providerId);
-    } catch {
-      console.error(`Unknown provider: ${providerId}`);
-      process.exitCode = 1;
-      return true;
-    }
-    saveOnboardingState(clearProviderOnboardingEntry(loadOnboardingState(), providerId));
-    if (!json) console.log(`Re-enabled ${providerId}.`);
-  }
-
-  const authStorage = AuthStorage.create();
-  const modelRegistry = ModelRegistry.create(authStorage);
-  const settingsManager = SettingsManager.create(cwd, agentDir);
-  const report = await buildDoctorReport({
-    cwd,
-    agentDir,
-    includeSessions: args.includes("--sessions"),
-    includeGui: args.includes("--full"),
-    modelSetup: buildCliModelSetupState(modelRegistry, settingsManager),
-  });
-  console.log(json ? JSON.stringify(report, null, 2) : renderDoctorReport(report));
-  return true;
-}
-
-function buildCliModelSetupState(
-  modelRegistry: ModelRegistry,
-  settingsManager: SettingsManager,
-): DoctorModelSetupState {
-  modelRegistry.refresh();
-  const provider = settingsManager.getDefaultProvider();
-  const modelId = settingsManager.getDefaultModel();
-  const activeModel = provider && modelId ? modelRegistry.find(provider, modelId) : undefined;
-  if (activeModel && modelRegistry.hasConfiguredAuth(activeModel)) {
-    return {
-      requirement: "ready",
-      currentModel: `${activeModel.provider}/${activeModel.id}`,
-    };
-  }
-
-  const availableModels = modelRegistry.getAvailable().map((model) => ({
-    provider: model.provider,
-    id: model.id,
-    label: `${model.provider}/${model.id}`,
-  }));
-  return {
-    requirement: availableModels.length > 0 ? "select_model" : "connect_auth",
-    currentModel: activeModel ? `${activeModel.provider}/${activeModel.id}` : undefined,
-    availableModels,
-  };
 }
 
 async function main(): Promise<void> {
